\name{dbgfns}
\alias{dbgfns}
\title{dbgfns}

\description{Sets debug/undebug status for the user's functions.}

\usage{
dbgfns(ALL=FALSE)
}

\arguments{
\item{ALL}{value TRUE sets debug status on all functions, while FALSE
will present user with menu of functions to toggle as to
debug/undebug status.}
}


\author{Norm Matloff}

