\name{dbgstep}
\alias{dbgstep}
\title{dbgstep}

\description{Updates cursor position at editor.}

\usage{
dbgstep()
}

\details{Is invoked from the editor to report to the latter the current
file and line position of the function being debugged, after an n or c
step in the browser.
}

\author{Norm Matloff}

